consolidate tests in a big file
