+++ soroban-sdk/src/tests.rs
@@ -4,8 +4,7 @@ mod budget;
 mod contract_add_i32;
 mod contract_assert;
 mod contract_call_stack;
-mod contract_docs_fn;
-mod contract_docs_struct;
+mod contract_docs;
 mod contract_invoke;
 mod contract_overlapping_type_fn_names;
 mod contract_snapshot;
@@ -17,7 +16,3 @@ mod contractfile_with_sha256;
 mod contractimport;
 mod contractimport_with_error;
 mod contractimport_with_sha256;
-mod contract_docs_struct_tuple;
-mod contract_docs_enum_int;
-mod contract_docs_enum_error_int;
-mod contract_docs_enum;
